Modularize hall of fame component
